+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Change Log
+## [Unreleased]
+### Changed
+- Switched to using native iRODS controllers, executables and helpers (no longer requires `test-with-irods`).
+- Groups predefined controllers by product (e.g. iRODS) not by type (e.g. service).
+- Allows quicker testing with latest versions of controllers only.
+
 ## 1.0.0
 ### Added
 - First stable release