+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Change Log
-## [Unreleased]
+## 2.0.0 - 2017-01-11
 ### Changed
 - Switched to using native iRODS controllers, executables and helpers (no longer requires `test-with-irods`).
 - Groups predefined controllers by product (e.g. iRODS) not by type (e.g. service).
@@ -7,6 +7,6 @@
 - Rebranded from `startfortest` to `useintest`.
 - Moved documentation to ReadTheDocs.
 
-## 1.0.0
+## 1.0.0 - 2016-11-23
 ### Added
 - First stable release